change GoalsSection to ProgressSection, dispatch action
