+++ src/js/components/home/HomePage.js
@@ -8,7 +8,7 @@ import ProofArticlesSection from './ProofArticlesSection';
 import EligibilityCheckCtaSection from './EligibilityCheckCtaSection';
 import MainBannerSection from './MainBannerSection';
 import VolunteerSection from './VolunteerSection';
-import GoalsSection from './GoalsSection';
+import ProgressSection from './ProgressSection';
 
 class HomePage extends React.Component {
   constructor(props, context){
@@ -69,7 +69,7 @@ class HomePage extends React.Component {
 
         <TestimonySection />
 
-        <GoalsSection />
+        <ProgressSection />
 
         <VideoSection />
